extract find or create short url to separate service file
